+++ src/nihm_index.h
@@ -106,7 +106,7 @@ class Index {
         query_output result;
         unordered_map<gid,uint32_t> counts;
         for(uint i(0);i<F;++i){
-            if(sketch[i]<fingerprint_range and sketch[i]>0){
+            if(sketch[i]<(int32_t)fingerprint_range and sketch[i]>0){
                 for(uint j(0);j<Buckets[sketch[i]+i*fingerprint_range].size();++j){
                     counts[Buckets[sketch[i]+i*fingerprint_range][j]]++;
                 }
@@ -127,28 +127,28 @@ class Index {
         if(W<=16){
             int16_t counts[genome_numbers];
             for(uint i(0);i<F;++i){
-                if(sketch[i]<fingerprint_range and sketch[i]>0){
+                if(sketch[i]<(int32_t)fingerprint_range and sketch[i]>0){
                     for(uint j(0);j<Buckets[sketch[i]+i*fingerprint_range].size();++j){
                         counts[Buckets[sketch[i]+i*fingerprint_range][j]]++;
                     }
                 }
             }
-            for(int32_t i(0);i<genome_numbers;++i){
-                if(counts[i]>=min_score){
+            for(uint32_t i(0);i<genome_numbers;++i){
+                if((uint32_t)counts[i]>=min_score){
                      result.push_back({counts[i],i});
                 }
             }
         }else{
             int32_t counts[genome_numbers];
             for(uint i(0);i<F;++i){
-                if(sketch[i]<fingerprint_range and sketch[i]>0){
+                if(sketch[i]<(int32_t)fingerprint_range and sketch[i]>0){
                     for(uint j(0);j<Buckets[sketch[i]+i*fingerprint_range].size();++j){
                         counts[Buckets[sketch[i]+i*fingerprint_range][j]]++;
                     }
                 }
             }
-            for(int32_t i(0);i<genome_numbers;++i){
-                if(counts[i]>=min_score){
+            for(uint32_t i(0);i<genome_numbers;++i){
+                if((uint32_t)counts[i]>=min_score){
                      result.push_back({counts[i],i});
                 }
             }
